move onCommand methods to tabExecutors
