+++ src/main/java/epicspellsplugin/MageManager.java
@@ -60,76 +60,6 @@ public class MageManager {
         }
     }
 
-    public boolean onCommand(CommandSender cs, Command cmd, String string, String[] args){
-        if(args.length > 0) {
-            Mage mage = getMage(args[0]);
-            if(mage != null) {
-                switch (args[1]) {
-                    case "mana":
-                        if(args.length == 4){
-                            try {
-                                float value = Float.valueOf(args[3]);
-                                switch (args[2]){
-                                    case "add":
-                                        mage.addMana(value);
-                                        break;
-                                    case "remove":
-                                        mage.removeMana(value);
-                                        break;
-                                    case "set":
-                                        mage.setMana(value);
-                                        break;
-                                }
-                            } catch (NumberFormatException e){
-                                cs.sendMessage(String.format("Error: %s is not a number!", args[3]));
-                            }
-                        } else if(args.length == 3){
-                            if(args[2].equals("get")){
-                                cs.sendMessage(String.format("Mana of %s is %.2f", args[0], mage.getMana()));
-                            }
-                        }
-                        break;
-                    case "maximumMana":
-                        if(args.length == 4){
-                            try {
-                                float value = Float.valueOf(args[3]);
-                                if(args[2].equals("set")){
-                                    mage.setMaxMana(value);
-                                }
-                            } catch (NumberFormatException e){
-                                cs.sendMessage(String.format("Error: %s is not a number!", args[3]));
-                            }
-                        } else if(args.length == 3) {
-                            if (args[2].equals("get")) {
-                                cs.sendMessage(String.format("Maximum Mana of %s is %.2f", args[0], mage.getMaxMana()));
-                            }
-                        }
-                        break;
-                    case "manaRegeneration":
-                        if(args.length == 4){
-                            try {
-                                float value = Float.valueOf(args[3]);
-                                if(args[2].equals("set")){
-                                    mage.setManaRegeneration(value);
-                                }
-                            } catch (NumberFormatException e){
-                                cs.sendMessage(String.format("Error: %s is not a number!", args[3]));
-                            }
-                        } else if(args.length == 3) {
-                            if (args[2].equals("get")) {
-                                cs.sendMessage(String.format("Mana regeneration rate of %s is %.2f", args[0], mage.getManaRegeneration()));
-                            }
-                        }
-                        break;
-                    case "remove":
-                        removePlayer(mage.getPlayer());
-                }
-            }
-            return true;
-        }
-        return false;
-    }
-
     public void castSpell(Mage mage, SpellWrapper spellWrapper){
         mage.removeMana(spellWrapper.getManaCost());
         mage.addCooldown(spellWrapper.getSpellName(), spellWrapper.getCooldown());
